==> scripts/1_cleanup_Data.ipynb <==

# CRICKET AND CODING - CRACK THE CHALLENGE 
## Problem Statement:

    Given certain input parameters regarding an innings of a T20 cricket match, predict the total runs scored by the batting team at the end of 6 overs.

## Input Data:
    the link to the Dataset (Source: cricsheet.org) which contains historic data of T20 matches that have occurred in the past. [https://cricsheet.org/downloads/ipl_male_csv2.zip] 



# 1. Download IPL Dataset

import requests, zipfile, io,glob,shutil,json
from pathlib import Path
from sklearn.preprocessing import LabelEncoder
from pprint import pprint
import numpy as np
import pandas as pd

data_path = Path(Path.cwd().parent,'dataset')

if not data_path.exists():
    # https://internalapp.nptel.ac.in/contest/reference_docs/ipl_csv2.zip
    # https://cricsheet.org/downloads/ipl_csv.zip
    # https://cricsheet.org/downloads/ipl_male_csv2.zip
    res = requests.get(r'https://internalapp.nptel.ac.in/contest/reference_docs/ipl_csv2.zip',stream=True)
    if res.status_code == 200:
        print('### Downloading the Dataset')
        z = zipfile.ZipFile(io.BytesIO(res.content))
        z.extractall(data_path)




df = pd.read_csv(Path.joinpath(data_path,'all_matches.csv'),parse_dates = ['start_date'])
#df_orig = df
print(df.columns)
df.head()

df.loc[df['start_date'].dt.year == 2021].groupby('batting_team')['striker'].unique().to_dict()

# 2. Clean the dataset
1. Filter the current seasion -2021 match details (which we use to test)
1. Drop columns :  'season','start_date'
2. Delete the non existing teams : 
        'Kochi Tuskers Kerala' 'Pune Warriors','Rising Pune Supergiants', 'Rising Pune Supergiant','Gujarat Lions'
        
3. Replace the OLD team names with new team name
4. Replace the same venue details with unique names 
5. Can remove the innings >2 details 


#df = df.loc[df['start_date'].dt.year < 2021]



# 1. Drop columns 
df = df.drop(columns=['season','start_date']) #,inplace=True)

# 2. Delete Non-existing teams : 'Kochi Tuskers Kerala' 'Pune Warriors','Rising Pune Supergiants', 'Rising Pune Supergiant','Gujarat Lions'

mask_bat_team = df['batting_team'].isin(['Kochi Tuskers Kerala',
                                'Pune Warriors',
                                'Rising Pune Supergiants',
                                'Rising Pune Supergiant',
                                'Gujarat Lions'                                
                            ])

mask_bow_team = df['bowling_team'].isin(['Kochi Tuskers Kerala',
                                'Pune Warriors',
                                'Rising Pune Supergiants',
                                'Rising Pune Supergiant',
                                'Gujarat Lions'                                
                            ])
df = df[~mask_bat_team]
df = df[~mask_bow_team]
print(df.shape)

#3. Replace the old team names with new team name:
df.loc[df.batting_team == 'Delhi Daredevils','batting_team'] = 'Delhi Capitals'
df.loc[df.batting_team == 'Deccan Chargers','batting_team'] = 'Sunrisers Hyderabad'

df.loc[df.bowling_team == 'Delhi Daredevils','bowling_team'] = 'Delhi Capitals'
df.loc[df.bowling_team == 'Deccan Chargers','bowling_team'] = 'Sunrisers Hyderabad'



# 4. Replace venue column unique names :
df.loc[df.venue == 'M.Chinnaswamy Stadium','venue'] = 'M Chinnaswamy Stadium'
df.loc[df.venue == 'Punjab Cricket Association IS Bindra Stadium, Mohali','venue'] = 'Punjab Cricket Association Stadium'
df.loc[df.venue == 'Punjab Cricket Association IS Bindra Stadium','venue'] = 'Punjab Cricket Association Stadium'
df.loc[df.venue == 'Wankhede Stadium, Mumbai','venue'] = 'Wankhede Stadium'
df.loc[df.venue == 'Rajiv Gandhi International Stadium, Uppal','venue'] = 'Rajiv Gandhi International Stadium'
df.loc[df.venue == 'MA Chidambaram Stadium, Chepauk' ,'venue'] = 'MA Chidambaram Stadium'
df.loc[df.venue == 'MA Chidambaram Stadium, Chepauk, Chennai' ,'venue'] = 'MA Chidambaram Stadium'

pprint('### Total {} : venue details present '.format(len(df.venue.unique())))
pprint('### Total {}  : Batting teams are there'.format(len(df.batting_team.unique())))
pprint('### Total {}  : Bowlling teams are there'.format(len(df.bowling_team.unique())))

print(df.shape)

#pd.DataFrame(df.groupby(['batting_team'])['striker'].unique())
df.groupby(['batting_team'])['bowler'].unique().to_dict()

list(df.venue.unique())

### Replace the team name in short

#df.loc[df.batting_team == 'Kolkata Knight Riders', 'batting_team'] = 'KKR'
#df.loc[df.batting_team == 'Royal Challengers Bangalore', 'batting_team'] = 'RCB'
#df.loc[df.batting_team == 'Chennai Super Kings', 'batting_team'] = 'CSK'
#df.loc[df.batting_team == 'Kings XI Punjab', 'batting_team'] = 'KXIP'
#df.loc[df.batting_team == 'Rajasthan Royals', 'batting_team'] = 'RR'
#df.loc[df.batting_team == 'Delhi Capitals', 'batting_team'] = 'DC'
#df.loc[df.batting_team == 'Sunrisers Hyderabad', 'batting_team'] = 'SRH'
#df.loc[df.batting_team == 'Mumbai Indians', 'batting_team'] = 'MI'

#df.loc[df.bowling_team == 'Kolkata Knight Riders', 'bowling_team'] = 'KKR'
#df.loc[df.bowling_team == 'Royal Challengers Bangalore', 'bowling_team'] = 'RCB'
#df.loc[df.bowling_team == 'Chennai Super Kings', 'bowling_team'] = 'CSK'
#df.loc[df.bowling_team == 'Kings XI Punjab', 'bowling_team'] = 'KXIP'
#df.loc[df.bowling_team == 'Rajasthan Royals', 'bowling_team'] = 'RR'
#df.loc[df.bowling_team == 'Delhi Capitals', 'bowling_team'] = 'DC'
#df.loc[df.bowling_team == 'Sunrisers Hyderabad', 'bowling_team'] = 'SRH'
#df.loc[df.bowling_team == 'Mumbai Indians', 'bowling_team'] = 'MI'

df.batting_team.unique()

# Rename striker and bowler column names:


df = df.rename(columns= {'striker' : 'batsmen','non_striker' : 'batsmen_non_striker','bowler' : 'bowlers'})
df.info()

df['Total_score']  = df.runs_off_bat + df.extras
df = df.drop(columns = ['wides','noballs', 'byes', 'legbyes', 'penalty', 'wicket_type', 'other_wicket_type', 'other_player_dismissed'],axis=1)
df[(df.ball<6.0) & (df.innings < 3)].to_csv('1_cleaned_Data.csv',index=False)


df = pd.read_csv('1_cleaned_Data.csv')
df.shape

players_df = pd.DataFrame(np.append(df.batsmen.unique(),df.bowlers.unique()),columns=['Players'])
players_df.head()

label_encode_dict = {}

le = LabelEncoder()
le.fit(df.batsmen)
batsmen_e = le.transform(df.batsmen)
batsmen_e_inv = le.inverse_transform(batsmen_e)

label_encode_dict['batsmen'] = dict(zip(batsmen_e_inv,map(int,batsmen_e)))

le.fit(df.bowlers)
bowlers_e = le.transform(df.bowlers)
bowlers_e_inv = le.inverse_transform(bowlers_e)

label_encode_dict['bowlers'] = dict(zip(bowlers_e_inv,map(int,bowlers_e)))

le.fit(df.venue)
venue_e = le.transform(df.venue)
venue_e_inv = le.inverse_transform(venue_e)

label_encode_dict['venue'] = dict(zip(venue_e_inv,map(int,venue_e)))


le.fit(df.batting_team)
batting_team_e = le.transform(df.batting_team)
batting_team_e_inv = le.inverse_transform(batting_team_e)

label_encode_dict['batting_team'] = dict(zip(batting_team_e_inv,map(int,batting_team_e)))

le.fit(df.bowling_team)
bowling_team_e = le.transform(df.bowling_team)
bowling_team_e_inv = le.inverse_transform(bowling_team_e)

label_encode_dict['bowling_team'] = dict(zip(bowling_team_e_inv,map(int,bowling_team_e)))



le = LabelEncoder()
le.fit(players_df.Players)
Players_e = le.transform(players_df.Players)
Players_e_inv = le.inverse_transform(Players_e)

label_encode_dict['Players'] = dict(zip(Players_e_inv,map(int,Players_e)))

with open(Path.joinpath(data_path,'label_encode.json'),'w') as f:
    json.dump(label_encode_dict,f)

players_df.Players.unique().shape

len(label_encode_dict['Players'])
