==> scripts/2_format_Data.ipynb <==
import requests, zipfile, io,glob,shutil
from pathlib import Path
from sklearn.preprocessing import LabelEncoder
from pprint import pprint
from collections import OrderedDict

import numpy as np
import pandas as pd

df = pd.read_csv('1_cleaned_Data.csv')


Runs_off_Bat_6_overs = df.groupby(['match_id','venue','innings','batting_team','bowling_team'])['runs_off_bat'].sum()
Extras_6_overs       = df.groupby(['match_id','venue','innings','batting_team','bowling_team'])['extras'].sum()
TotalScore_6_overs   = df.groupby(['match_id','venue','innings','batting_team','bowling_team'])['Total_score'].sum()

Total_WktsDown      = df.groupby(['match_id','venue','innings','batting_team','bowling_team'])['player_dismissed'].count()

bat_df   = df.groupby(['match_id','venue','innings','batting_team','bowling_team'])['batsmen'].apply(list)
bow_df   = df.groupby(['match_id','venue','innings','batting_team','bowling_team'])['bowlers'].apply(list)

pd.concat([bat_df,bow_df,Runs_off_Bat_6_overs,Extras_6_overs,TotalScore_6_overs,Total_WktsDown],axis=1).to_csv('test.csv')

df = pd.read_csv('test.csv')
df.head()

bat  = pd.DataFrame(np.zeros((df.shape[0],10),dtype=float),columns=['bat1', 'bat2', 'bat3', 'bat4', 'bat5', 'bat6', 'bat7', 'bat8', 'bat9', 'bat10'])
bowl = pd.DataFrame(np.zeros((df.shape[0],6),dtype=float),columns=['bow1', 'bow2', 'bow3', 'bow4', 'bow5', 'bow6'])

columns = ['bat1', 'bat2', 'bat3', 'bat4', 'bat5', 'bat6', 'bat7', 'bat8', 'bat9', 'bat10','bow1', 'bow2', 'bow3', 'bow4', 'bow5', 'bow6']
df_bat_bow = pd.concat([bat,bowl],axis=1)
df_bat_bow



pd.concat([df,df_bat_bow],axis=1).to_csv('test.csv',index=False)

df = pd.read_csv('test.csv')
os.remove('test.csv')
df.head()

for row,val in enumerate(df.batsmen):
    tmp = val[1:-1].split(', ')
    tmp_list = list(map(lambda x : x.strip("'"),tmp))
    tmp_list = list(OrderedDict.fromkeys(tmp_list))
    for i,j in enumerate(list(map(lambda x : x.strip("'"),tmp_list))):
        col = "bat%i"%(i+1)
        df[col][row] = j

for row,val in enumerate(df.bowlers):
    tmp = val[1:-1].split(', ')
    tmp_list = list(map(lambda x : x.strip("'"),tmp))
    tmp_list = list(OrderedDict.fromkeys(tmp_list))
    for i,j in enumerate(list(map(lambda x : x.strip("'"),tmp_list))):
        col = "bow%i"%(i+1)
        df[col][row] = j

df.to_csv('2_formated_Dataset.csv',index=False)

df = pd.read_csv('2_formated_Dataset.csv')


df.head()
