==> scripts/3_model_sklearn.ipynb <==
import pandas as pd
import numpy as np
import json

from matplotlib import pyplot as plt
%matplotlib inline

from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn import metrics

df = pd.read_csv('2_formated_Dataset.csv')
df.columns

df_model = df[['venue', 'innings', 'batting_team', 'bowling_team', 'bat1', 'bat2', 'bat3', 'bat4', 'bat5', 'bat6','bat7', 'bat8', 'bat9', 'bat10', 'bow1', 'bow2', 'bow3', 'bow4', 'bow5','bow6','runs_off_bat', 'extras', 'Total_score','player_dismissed']]

#df_model.to_csv('3_dataset.csv',index=False)

df_model.columns

# Encode the Labels from the existing encoded values

with open('label_encode.json') as f:
    data = json.load(f)

condition = False

for col in df_model.columns:
    if col in data.keys():
        condition = True
        col = col
    elif col in ['bat1', 'bat2',
       'bat3', 'bat4', 'bat5', 'bat6', 'bat7', 'bat8', 'bat9', 'bat10']:
        condition = True
        col = 'batsmen'
    elif col in ['bow1',
       'bow2', 'bow3', 'bow4', 'bow5', 'bow6']:
        col = 'bowlers'
        condition = True

    if condition:
        condition = False
        for key in data[col]:
            df_model = df_model.replace([key],data[col][key])


df_model.to_csv('3_dataset.csv',index=False)

df_model.columns

y = df_model['Total_score']
x = df_model[['venue', 'innings', 'batting_team', 'bowling_team', 'bat1', 'bat2',
       'bat3', 'bat4', 'bat5', 'bat6', 'bat7', 'bat8', 'bat9', 'bat10', 'bow1',
       'bow2', 'bow3', 'bow4', 'bow5', 'bow6']]

x_train, x_test, y_train, y_test = train_test_split(x,y,test_size=0.3,shuffle=True)

y_test_df = pd.DataFrame(y_test)
y_test_df

print(type(x_test))
print(type(y_test_df))

from sklearn.linear_model import LinearRegression

model = LinearRegression()
model.fit(x_train, y_train)

print(model.coef_)
print('\n\n\n')
print(model.intercept_)

pd.DataFrame(model.coef_, x.columns, columns = ['Coeff'])

predictions = model.predict(x_test)
df_pred = pd.DataFrame(predictions,columns=['Predicted_Total_score'])
df_pred

print(type(x_test))
print(type(y_test_df))
print(type(df_pred))



print(x_test.shape)
print(y_test_df.shape)
print(df_pred.shape)

x_test

y_test_df

df_pred

df = pd.concat([x_test,y_test_df,df_pred],axis=1).dropna()
df.head()

df.to_csv('4_Predicted_Results.csv',index=False)

plt.scatter(y_test_df, df_pred)

plt.hist(y_test - predictions)

metrics.mean_absolute_error(y_test, predictions)

metrics.mean_squared_error(y_test, predictions)

np.sqrt(metrics.mean_squared_error(y_test, predictions))
