==> scripts/4_model_pytorch.ipynb <==
import pandas as pd
import numpy as np
import torch

from sklearn.preprocessing import MinMaxScaler

import torch.nn as nn
from torch.utils.data import TensorDataset
from torch.utils.data import DataLoader

# Import nn.functional
import torch.nn.functional as F


df = pd.read_csv('3_dataset.csv')
print(df.shape)
df.head()

inputs_df = df[['venue','innings','batting_team', 'bowling_team', 'bat1', 'bat2','bat3', 'bat4', 'bat5', 'bat6', 'bat7', 'bat8', 'bat9', 'bat10', 'bow1','bow2', 'bow3', 'bow4', 'bow5', 'bow6']]#.to_numpy() #dtype=float)

targets_df = df[['Total_score']]#.to_numpy()#dtype=float)

print(inputs_df.shape,targets_df.shape)
inputs_df.head()



targets_df.head()

# Normalize data:


mms_input_obj = MinMaxScaler()
df_input_norm = pd.DataFrame(mms_input_obj.fit_transform(inputs_df),columns=inputs_df.columns)
df_input_norm.head()

mms_output_obj = MinMaxScaler()
df_output_norm = pd.DataFrame(mms_output_obj.fit_transform(targets_df),columns=targets_df.columns)
df_output_norm.head()

inputs_np = df_input_norm.to_numpy()
targets_np = df_output_norm.to_numpy()

inputs = torch.from_numpy(inputs_np).float()
targets = torch.from_numpy(targets_np).float()
targets = targets.view(1401,1)
print(inputs)
print(targets)

print(inputs.shape)
print(targets.shape)

# Define dataset
train_ds = TensorDataset(inputs, targets)
len(train_ds)

# Define data loader
batch_size = 5
train_dl = DataLoader(train_ds, batch_size, shuffle=True)


for i,j in train_dl:
    print(i.shape,j.shape)
    print(i)
    print(j)
    break

for xb, yb in train_dl:
    print(xb)
    print(yb)
    break

#torch.autograd.detect_anomaly(True)

# Define model
model = nn.Linear(20,1)
print(model.weight.shape)
print(model.bias.shape)

# Parameters
list(model.parameters())

# Generate predictions
preds = model(inputs)
preds

# Loss Function


# Define loss function
loss_fn = nn.MSELoss()

loss = loss_fn(model(inputs), targets)
print(loss)

# Optimizer

# Define optimizer
opt = torch.optim.SGD(model.parameters(), lr=1e-1)

# Train the model

We are now ready to train the model. We'll follow the same process to implement gradient descent:

1. Generate predictions
2. Calculate the loss
3. Compute gradients w.r.t the weights and biases
4. Adjust the weights by subtracting a small quantity proportional to the gradient
5. Reset the gradients to zero

The only change is that we'll work batches of data instead of processing the entire training data in every iteration. Let's define a utility function fit that trains the model for a given number of epochs.

# Utility function to train the model
def fit(num_epochs, model, loss_fn, opt, train_dl):
    # Repeat for given number of epochs
    for epoch in range(num_epochs):
        # Train with batches of data
        for xb,yb in train_dl:
            # 1. Generate predictions
            pred = model(xb)
            #print(pred)
            #import pdb;pdb.set_trace()
            #print(pred);import pdb;pdb.set_trace()
            # 2. Calculate loss
            loss = loss_fn(pred, yb)
            # 3. Compute gradients
            loss.backward()
            # 4. Update parameters using gradients
            opt.step()
            # 5. Reset the gradients to zero
            opt.zero_grad()
        # Print the progress
        if (epoch+1) % 10 == 0:
            print('Epoch [{}/{}], Loss: {}'.format(epoch+1, num_epochs, loss.item()))

fit(100, model, loss_fn, opt, train_dl)

# Generate predictions
preds = model(inputs)
preds

#pd.DataFrame(,index=False,columns='Actual Score in 6-overs')
mms_output_obj.inverse_transform(targets)

pred_df = pd.DataFrame(mms_output_obj.inverse_transform(preds.detach().numpy()),columns=['Predicted Score in 6-Overs'])

pd.concat([inputs_df,targets_df,pred_df],axis=1).to_csv('model_pred.csv')
